Add New Colmap-Wrapper

+++ src/main/python/render/data/GeometryO3D.py
@@ -35,19 +35,19 @@ class GeometryO3DTriangleMesh(Geometry):
         self.primtive = Primitves.TRIANGLES
         
         if self.triangle_mesh.has_vertices():
-            self.vertices: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertices).astype('float32'), PrimitiveType.DOUBLE)
+            self.vertices: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertices).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.vertices)
         
         if self.triangle_mesh.has_triangle_normals():
-            self.normales: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertex_normals).astype('float32'), PrimitiveType.DOUBLE)
+            self.normales: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertex_normals).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.normales)
         
         if self.triangle_mesh.has_vertex_colors():
-            self.colors: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertex_colors).astype('float32'), PrimitiveType.DOUBLE)
+            self.colors: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.vertex_colors).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.colors)
         
         if self.triangle_mesh.has_triangle_uvs():
-            self.uvs: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.triangle_uvs).astype('float32'), PrimitiveType.DOUBLE)
+            self.uvs: GeometryData = GeometryData(3, np.asarray(self.triangle_mesh.triangle_uvs).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.uvs)
 
 class GeometryO3DLineSet(Geometry):
@@ -60,9 +60,9 @@ class GeometryO3DLineSet(Geometry):
         # TODO: has_points
         
         if self.line_set.has_lines():
-            self.vertices: GeometryData = GeometryData(3, np.asarray(self.line_set.lines).astype('float32'), PrimitiveType.DOUBLE)
+            self.vertices: GeometryData = GeometryData(3, np.asarray(self.line_set.lines).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.vertices)
         
         if self.line_set.has_colors():
-            self.colors: GeometryData = GeometryData(3, np.asarray(self.line_set.colors).astype('float32'), PrimitiveType.DOUBLE)
+            self.colors: GeometryData = GeometryData(3, np.asarray(self.line_set.colors).astype('float32'), PrimitiveType.FLOAT)
             self.all_vertices.append(self.colors)
